% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{cdfij}
\alias{cdfij}
\title{Cumulative distribution function of a distributional data unit}
\usage{
cdfij(obj, i, j)
}
\arguments{
\item{obj}{an object of class \code{ddojb}}

\item{i}{unit number}

\item{j}{variable number}
}
\value{
a function(x) that will evaluate the c.d.f. at the values in x
}
\description{
Cumulative distribution function of a distributional data unit
}
