% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{create_ddobj}
\alias{create_ddobj}
\title{Creates a distributional data object}
\usage{
create_ddobj(df, types=NULL, cols=NULL, n.int=NULL, n.cat=NULL)
}
\arguments{
\item{df}{data frame from which the \code{ddobj} object is created}

\item{types}{types of distributional variables provided as a vector. Possible types are
\code{numeric}, \code{interval}, \code{histogram}, \code{categorical}, \code{modal}}

\item{cols}{the first column number in \code{df} corresponding to each of the `types`}

\item{n.int}{number of intervals for each histogram variable}

\item{n.cat}{number of categories for each modal variable}
}
\value{
an object of class \code{ddobj}
}
\description{
Creates a distributional data object
}
\examples{
create_ddobj (toy.data, types=c("numeric","interval","histogram","categorical", "modal"),
              cols = c(1, 2, 4, 13, 14), n.int=4, n.cat=3)

}
