% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{Creditcard.data}
\alias{Creditcard.data}
\title{Credit card data.}
\format{
An object of class \code{ddobj}. A list of length 5:
\describe{
  \item{Food}{interval scaled variable}
  \item{Social}{interval scaled variable}
  \item{Travel}{interval scaled variable}
  \item{Gas}{interval scaled variable}
  \item{Clothes}{numeric variable}
  }
}
\usage{
Creditcard.data
}
\description{
A data set containing 10 units with five interval scaled variables.
}
\keyword{datasets}
