% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{dataSDAhist_to_cols}
\alias{dataSDAhist_to_cols}
\title{Convert histogram-valued data from dataSDA to data.frame}
\usage{
dataSDAhist_to_cols(hist_mat)
}
\arguments{
\item{hist_mat}{output of the \code{hist_extract()} function}
}
\value{
a data frame for input to create_ddobj
}
\description{
Convert histogram-valued data from dataSDA to data.frame
}
