% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{dataSDAmodal_to_cols}
\alias{dataSDAmodal_to_cols}
\title{Convert modal-valued data from dataSDA to data.frame}
\usage{
dataSDAmodal_to_cols(modal_mat)
}
\arguments{
\item{modal_mat}{output of the \code{hist_extract()} function}
}
\value{
a data frame for input to create_ddobj
}
\description{
Convert modal-valued data from dataSDA to data.frame
}
