% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{ddbiplot}
\alias{ddbiplot}
\title{First step to create a new biplot with \pkg{biplotEZ}}
\usage{
ddbiplot(data, classes = NULL, group.aes = NULL, center = TRUE,
                scaled = c(NULL, "Wasserstein", "Billard"), Title = NULL)
}
\arguments{
\item{data}{an object of class \code{ddobj}.}

\item{classes}{a vector identifying class membership.}

\item{group.aes}{a vector identifying groups for aesthetic formatting.}

\item{center}{a logical value indicating whether \code{data} should be variable centered,
with default \code{TRUE}.}

\item{scaled}{default \code{NULL} for no scaling of the data. Other options are
\code{Billard} or \code{Wasserstein} to divide centred quantile functions
by the squareroot of the variance computed with \code{ddvarB} or
\code{ddvarW} respectively. Partial matching allowed.}

\item{Title}{the title of the biplot to be rendered, enter text in "  ".}
}
\value{
A list with the following components is available:
\item{X}{numeric part of the original \code{ddobj}, centred and scaled according to the
         arguments \code{center} and \code{scaled}.}
\item{Xcat}{categorical and modal part of the original \code{ddobj}.}
\item{raw.X}{original \code{ddobj}.}
\item{numeric.vars}{which of the variables in raw.X are contained in X.}
\item{classes}{the vector of category levels for the class variable. This is to be used for
               \code{colour}, \code{pch} and \code{cex} specifications.}
\item{na.action}{the observations that have been removed.}
\item{center}{a logical value indicating whether centering was applied.}
\item{scaled}{one of \code{NULL} or \code{Billard} or \code{Wasserstein}.}
\item{means}{the vector of means for each numeric, interval or histogram variable.}
\item{sd}{the vector of standard deviations computed according to the specification in\code{scaled}.}
\item{n}{the number of observations.}
\item{p}{the number of variables.}
\item{group.aes}{the vector of category levels for the grouping variable. This is to be used
                 for \code{colour}, \code{pch} and \code{cex} specifications.}
\item{g.names}{the descriptive names to be used for group labels.}
\item{g}{the number of groups.}
\item{Title}{the title of the biplot rendered}
}
\description{
This function produces a \code{ddbiplot} object.
}
\details{
This function is the entry-level function in \code{DDbiplotEZ} to construct a biplot display.
It initialises an object of class \code{ddbiplot} which can then be piped to various other functions
to build up the biplot display.
}
\examples{
ddbiplot(data = Oils.data)
# create a PCA biplot
ddbiplot(data = Oils.data) |> PCA() |> plot()

}
