% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{ddcor}
\alias{ddcor}
\title{Computes the correlation of two distributional data objects}
\usage{
ddcor(x, y)
}
\arguments{
\item{x}{a list, typically a single component of an object of class \code{ddojb}}

\item{y}{a list, typically a single component of an object of class \code{ddojb}}
}
\value{
numeric correlation value
}
\description{
Computes the correlation of two distributional data objects
}
\examples{
obj <- suminto.ddobj (esoph, units = "agegp", interval="ncases", histogram="ncontrols")
ddcor (obj$ncases, obj$ncontrols)
}
