% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{ddcorB}
\alias{ddcorB}
\title{Computes the correlation of two distributional data objects as defined by Billard (2008)}
\usage{
ddcorB(x, y)
}
\arguments{
\item{x}{a list, typically a single component of an object of class \code{ddojb}}

\item{y}{a list, typically a single component of an object of class \code{ddojb}}
}
\value{
numeric correlation value
}
\description{
Computes the correlation of two distributional data objects as defined by Billard (2008)
}
\examples{
obj <- suminto_ddobj (esoph, units = "agegp", interval="ncases", histogram="ncontrols")
ddcorB (obj$ncases, obj$ncontrols)
}
