% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{ddcorW}
\alias{ddcorW}
\title{Computes the correlation based on Wasserstein distance of distributional data objects}
\usage{
ddcorW(x)
}
\arguments{
\item{x}{an object of class \code{ddojb}}
}
\value{
correlation matrix
}
\description{
This function computes correlations from the variances and covariances based on the
\eqn{L_2} Wasserstein distance as defined in Irpino and Verde (2015).
}
\examples{
obj <- suminto_ddobj (esoph, units = "agegp", interval="ncases", histogram="ncontrols")
ddcorW (obj)

}
\references{
Irpino, A. and Verde, R. 2015. Basic statistics for distributional symbolic variables:
a new metric-based approach. Advances in Data Analysis and Classification, 9(2), pp.143-157.
}
