% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{ddcovB}
\alias{ddcovB}
\title{Computes the covariance of two distributional data objects as defined by Billard (2008)}
\usage{
ddcovB(x, y)
}
\arguments{
\item{x}{a list, typically a single component of an object of class \code{ddobj}}

\item{y}{a list, typically a single component of an object of class \code{ddobj}}
}
\value{
numeric variance or covariance value
}
\description{
This function only works for an interval scaled or histogram scaled variable. For an interval
scaled variable \code{x} should be a list with component \code{$type = "interval"} and
component \code{values} a two-column matrix of interval endpoints. If \code{x} is a histogram
scaled variable, it should be a list with component \code{$type = "histogram"} and components
\code{intervals} and \code{proportions}.
}
\examples{
obj <- suminto_ddobj (esoph, units = "agegp", interval="ncases", histogram="ncontrols")
ddcovB (obj$ncases, obj$ncontrols)

}
\references{
Billard, L., 2008. Sample covariance functions for complex quantitative data. In Proceedings
of IASC 2008, Yokohama, Japan, pp 157-163.
}
