% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{ddcovmat}
\alias{ddcovmat}
\title{Computes the covariance matrix of distributional data variables}
\usage{
ddcovmat(obj)
}
\arguments{
\item{obj}{an object of class \code{ddobj}}
}
\value{
a covariance matrix of size length of obj by length of obj
}
\description{
Computes the covariance matrix of distributional data variables
}
\examples{
obj <- suminto.ddobj (esoph, units = "agegp", interval="ncases", histogram="ncontrols")
ddcovmat (obj)
}
