% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{ddcovmatB}
\alias{ddcovmatB}
\title{Computes the covariance matrix of distributional data variables as defined by Billard (2008)}
\usage{
ddcovmatB(obj)
}
\arguments{
\item{obj}{an object of class \code{ddobj}}
}
\value{
a covariance matrix of size length of obj by length of obj
}
\description{
Computes the covariance matrix of distributional data variables as defined by Billard (2008)
}
\examples{
obj <- suminto_ddobj (esoph, units = "agegp", interval="ncases", histogram="ncontrols")
ddcovmatB (obj)
}
