% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{ddmean}
\alias{ddmean}
\title{Computes the mean of a distributional data object}
\usage{
ddmean(x)
}
\arguments{
\item{x}{a list, typically a single component of an object of class \code{ddojb}}
}
\value{
numeric mean value
}
\description{
This function only works for an interval scaled or histogram scaled variable. For an interval
scaled variable \code{x} should be a list with component \code{$type = "interval"} and
component \code{values} a two-column matrix of interval endpoints. If \code{x} is a histogram
scaled variable, it should be a list with component \code{$type = "histogram"} and components
\code{intervals} and \code{proportions}.
}
\examples{
obj <- suminto.ddobj (esoph, units = "agegp", interval="ncases", histogram="ncontrols")
ddmean (obj$ncases)
ddmean (obj$ncontrols)

}
\references{
Billard, L., 2008. Sample covariance functions for complex quantitative
 data. In Proceedings of IASC 2008, Yokohama, Japan, pp 157-163.
}
