% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{ddQmean}
\alias{ddQmean}
\title{Computes the Frechet mean's quantile function}
\usage{
ddQmean(x, j)
}
\arguments{
\item{x}{an object of class \code{ddobj}}

\item{j}{the number(s) of the variable(s) for which the quantile function is computed.
If \code{j} is null, the computation is performed for all variables.}
}
\value{
a the quantile function
}
\description{
Computes the Frechet mean's quantile function
}
\examples{
obj <- suminto_ddobj (esoph, units = "agegp", interval="ncases", histogram="ncontrols")
ddQmean (obj, j=1)(seq(from=0, to=1, len=20))

}
