% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{ddvarB}
\alias{ddvarB}
\title{Computes the variance or covariance of (a) distributional data object(s) as defined by Billard (2008)}
\usage{
ddvarB(x, y)
}
\arguments{
\item{x}{a list, typically a single component of an object of class \code{ddojb}}

\item{y}{optional, a list, typically a single component of an object of class \code{ddojb}. If
\code{y} is specified, the covariance is computed.}
}
\value{
numeric variance or covariance value
}
\description{
This function only works for an interval scaled or histogram scaled variable. For an interval
scaled variable \code{x} should be a list with component \code{$type = "interval"} and
component \code{values} a two-column matrix of interval endpoints. If \code{x} is a histogram
scaled variable, it should be a list with component \code{$type = "histogram"} and components
\code{intervals} and \code{proportions}.
}
\examples{
obj <- suminto_ddobj (esoph, units = "agegp", interval="ncases", histogram="ncontrols")
ddvarB (obj$ncases) # variance
ddvarB (obj$ncases, obj$ncontrols) # covariance

}
\references{
Billard, L., 2008. Sample covariance functions for complex quantitative data. In Proceedings
of IASC 2008, Yokohama, Japan, pp 157-163.
}
