% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{ddvarW}
\alias{ddvarW}
\title{Computes the variance or covariance based on Wasserstein distance of
(a) distributional data object(s)}
\usage{
ddvarW(x)
}
\arguments{
\item{x}{an object of class \code{ddojb}}
}
\value{
variance-covariance matrix
}
\description{
This function computes variances and covariances based on the \eqn{L_2} Wasserstein
distance as defined in Irpino and Verde (2015).
}
\examples{
obj <- suminto_ddobj (esoph, units = "agegp", interval="ncases", histogram="ncontrols")
ddvarW (obj)

}
\references{
Irpino, A. and Verde, R. 2015. Basic statistics for distributional symbolic variables:
a new metric-based approach. Advances in Data Analysis and Classification, 9(2), pp.143-157.
}
