% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{extractDDunits}
\alias{extractDDunits}
\title{Extract some units from a \code{ddobj}}
\usage{
extractDDunits(obj, i)
}
\arguments{
\item{obj}{an object of class \code{ddobj}}

\item{i}{a vector of the numbers of the units to extract}
}
\value{
an object of class \code{ddobj}
}
\description{
Extract some units from a \code{ddobj}
}
\examples{
my.obj <- create_ddobj (toy.data, type=c("numeric","interval","histogram","categorical", "modal"),
          cols = c(1, 2, 4, 13, 14), n.int=4, n.cat=3)
extractDDunits (my.obj, i=c(1,3:4))

}
