% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{extractDDvars}
\alias{extractDDvars}
\title{Extract some variables from a \code{ddobj}}
\usage{
extractDDvars(obj, j)
}
\arguments{
\item{obj}{an object of class \code{ddobj}}

\item{j}{a vector of the numbers of the variables to extract}
}
\value{
an object of class \code{ddobj}
}
\description{
Extract some variables from a \code{ddobj}
}
\examples{
my.obj <- create_ddobj (toy.data, type=c("numeric","interval","histogram","categorical", "modal"),
          cols = c(1, 2, 4, 13, 14), n.int=4, n.cat=3)
extractDDvars (my.obj, j=c(1,3))

}
