% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.abalone.int}
\alias{get.abalone.int}
\title{Converts the abalone.int data set to a ddobj}
\usage{
get.abalone.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
abalone.data <- get.abalone.int()

}
