% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.Abalone}
\alias{get.Abalone}
\title{Converts the Abalone data set to a ddobj}
\usage{
get.Abalone()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
Abalone.data <- get.Abalone()

}
