% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.acid_rain.int}
\alias{get.acid_rain.int}
\title{Converts the acid_rain.int data set to a ddobj}
\usage{
get.acid_rain.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
acidRain.data <- get.acid_rain.int()

}
