% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.age_cholesterol_weight.int}
\alias{get.age_cholesterol_weight.int}
\title{Converts the age_choloesterol_weight.int data set to a ddobj}
\usage{
get.age_cholesterol_weight.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
age.chol.wt.data <- get.age_cholesterol_weight.int()

}
