% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.age_pyramids.hist}
\alias{get.age_pyramids.hist}
\title{Converts the age_pyramids.hist data set to a ddobj}
\usage{
get.age_pyramids.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
pyramids.data <- get.age_pyramids.hist()

}
