% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.Age_Pyramids_2014}
\alias{get.Age_Pyramids_2014}
\title{Converts the Age_Pyramids_2014 data set to a ddobj}
\usage{
get.Age_Pyramids_2014()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{HistDAWass} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
pyramid.data <- get.Age_Pyramids_2014()

}
