% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.Agronomique}
\alias{get.Agronomique}
\title{Converts the Agronomique data set to a ddobj}
\usage{
get.Agronomique()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{HistDAWass} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
pyramid.data <- get.Agronomique()

}
