% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.airline_flights2.modal}
\alias{get.airline_flights2.modal}
\title{Converts the airline_flights2.modal data set to a ddobj}
\usage{
get.airline_flights2.modal()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
flights.data <- get.airline_flights2.modal()

}
