% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.baseball.int}
\alias{get.baseball.int}
\title{Converts the baseball.int data set to a ddobj}
\usage{
get.baseball.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
baseball.data <- get.baseball.int()

}
