% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.bats.int}
\alias{get.bats.int}
\title{Converts the bats.int data set to a ddobj}
\usage{
get.bats.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
bats.data <- get.bats.int()

}
