% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.bird_species_extended.mix}
\alias{get.bird_species_extended.mix}
\title{Converts the bird_species_extended.mix data set to a ddobj}
\usage{
get.bird_species_extended.mix()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
bird.data <- get.bird_species_extended.mix()

}
