% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.blood.hist}
\alias{get.blood.hist}
\title{Converts the blood.hist data set to a ddobj}
\usage{
get.blood.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
blood.data <- get.blood.hist()

}
