% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.blood_pressure.int}
\alias{get.blood_pressure.int}
\title{Converts the blood_pressure.int data set to a ddobj}
\usage{
get.blood_pressure.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
bp.data <- get.blood_pressure.int()

}
