% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.BloodBRITO}
\alias{get.BloodBRITO}
\title{Converts the BloodBRITO data set to a ddobj}
\usage{
get.BloodBRITO()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{HistDAWass} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
blood.data <- get.BloodBRITO()

}
