% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.car_models.int}
\alias{get.car_models.int}
\title{Converts the car_models.int data set to a ddobj}
\usage{
get.car_models.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
car_models.data <- get.car_models.int()

}
