% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.cardiological.int}
\alias{get.cardiological.int}
\title{Converts the cardiological.int data set to a ddobj}
\usage{
get.cardiological.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
cardiological.data <- get.cardiological.int()

}
