% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.cars.int}
\alias{get.cars.int}
\title{Converts the cars.int data set to a ddobj}
\usage{
get.cars.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
cars.data <- get.cars.int()

}
