% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.Cars}
\alias{get.Cars}
\title{Converts the Cars data set to a ddobj}
\usage{
get.Cars()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{MAINT.Data} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
Cars.data <- get.Cars()

}
