% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.census.mix}
\alias{get.census.mix}
\title{Converts the census.mix data set to a ddobj}
\usage{
get.census.mix()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
census.data <- get.census.mix()

}
