% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.china_climate_month.hist}
\alias{get.china_climate_month.hist}
\title{Converts the china_climate_month.hist data set to a ddobj}
\usage{
get.china_climate_month.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
China.data <- get.china_climate_month.hist()

}
