% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.China_Seas}
\alias{get.China_Seas}
\title{Converts the China_Seas data set to a ddobj}
\usage{
get.China_Seas()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{HistDAWass} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
blood.data <- get.China_Seas()

}
