% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.china_temp_monthly.int}
\alias{get.china_temp_monthly.int}
\title{Converts the china_temp_monthly.int data set to a ddobj}
\usage{
get.china_temp_monthly.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
ChinaTemp.data <- get.china_temp_monthly.int()

}
