% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.ChinaTemp}
\alias{get.ChinaTemp}
\title{Converts the ChinaTemp data set to a ddobj}
\usage{
get.ChinaTemp()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{MAINT.Data} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
China.data <- get.ChinaTemp()

}
