% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.cholesterol.hist}
\alias{get.cholesterol.hist}
\title{Converts the cholesterol.hist data set to a ddobj}
\usage{
get.cholesterol.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
cholesterol.data <- get.cholesterol.hist()

}
