% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.county_income_gender.hist}
\alias{get.county_income_gender.hist}
\title{Converts the county_income_gender.hist data set to a ddobj}
\usage{
get.county_income_gender.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
county.data <- get.county_income_gender.hist()

}
