% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.cover_types.hist}
\alias{get.cover_types.hist}
\title{Converts the cover_types.hist data set to a ddobj}
\usage{
get.cover_types.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
cover.data <- get.cover_types.hist()

}
