% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.credit_card.int}
\alias{get.credit_card.int}
\title{Converts the credit_card.int data set to a ddobj}
\usage{
get.credit_card.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
creditCard.data <- get.credit_card.int()

}
