% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.crime.modal}
\alias{get.crime.modal}
\title{Converts the crime.modal data set to a ddobj}
\usage{
get.crime.modal()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
crime.data <- get.crime.modal()

}
