% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.ecoli_routes.int}
\alias{get.ecoli_routes.int}
\title{Converts the ecoli_routes.int data set to a ddobj}
\usage{
get.ecoli_routes.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
ecoliRoutes.data <- get.ecoli_routes.int()

}
