% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.employment.int}
\alias{get.employment.int}
\title{Converts the employment.int data set to a ddobj}
\usage{
get.employment.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
employment.data <- get.employment.int()

}
