% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.environment.mix}
\alias{get.environment.mix}
\title{Converts the environment.mix data set to a ddobj}
\usage{
get.environment.mix()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
environment.data <- get.environment.mix()

}
