% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.exchange_rate_returns.hist}
\alias{get.exchange_rate_returns.hist}
\title{Converts the exchange_rate_returns.hist data set to a ddobj}
\usage{
get.exchange_rate_returns.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
returns.data <- get.exchange_rate_returns.hist()

}
