% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.finance.int}
\alias{get.finance.int}
\title{Converts the finance.int data set to a ddobj}
\usage{
get.finance.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
fin.data <- get.finance.int()

}
