% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.flights_detail.hist}
\alias{get.flights_detail.hist}
\title{Converts the flights_detail.hist data set to a ddobj}
\usage{
get.flights_detail.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
flights.data <- get.flights_detail.hist()

}
