% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.french_agriculture.hist}
\alias{get.french_agriculture.hist}
\title{Converts the french_agriculture.hist data set to a ddobj}
\usage{
get.french_agriculture.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
agri.data <- get.french_agriculture.hist()

}
