% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.freshwater_fish.int}
\alias{get.freshwater_fish.int}
\title{Converts the freshwater_fish.int data set to a ddobj}
\usage{
get.freshwater_fish.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
fish.data <- get.freshwater_fish.int()

}
