% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.fuel_consumption.modal}
\alias{get.fuel_consumption.modal}
\title{Converts the fuel_consumption.modal data set to a ddobj}
\usage{
get.fuel_consumption.modal()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
fuel.data <- get.fuel_consumption.modal()

}
