% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.fungi.int}
\alias{get.fungi.int}
\title{Converts the fungi.int data set to a ddobj}
\usage{
get.fungi.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
fungi.data <- get.fungi.int()

}
