% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.genome_abundances.int}
\alias{get.genome_abundances.int}
\title{Converts the genome_abundances.int data set to a ddobj}
\usage{
get.genome_abundances.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
genome.data <- get.genome_abundances.int()

}
