% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.glucose.hist}
\alias{get.glucose.hist}
\title{Converts the glucose.hist data set to a ddobj}
\usage{
get.glucose.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
glucose.data <- get.glucose.hist()

}
