% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.hardwood.hist}
\alias{get.hardwood.hist}
\title{Converts the hardwood.hist data set to a ddobj}
\usage{
get.hardwood.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
hardwood.data <- get.hardwood.hist()

}
