% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.hdi_gender.int}
\alias{get.hdi_gender.int}
\title{Converts the hdi_gender.int data set to a ddobj}
\usage{
get.hdi_gender.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
HDIgender.data <- get.hdi_gender.int()

}
