% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.hematocrit.hist}
\alias{get.hematocrit.hist}
\title{Converts the hematocrit.hist data set to a ddobj}
\usage{
get.hematocrit.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
hematocrit.data <- get.hematocrit.hist()

}
