% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.hematocrit_hemoglobin.hist}
\alias{get.hematocrit_hemoglobin.hist}
\title{Converts the hematocrit_hemoglobin.hist data set to a ddobj}
\usage{
get.hematocrit_hemoglobin.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
hematocrit_hemoglobin.data <- get.hematocrit_hemoglobin.hist()

}
