% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.hierarchy.hist}
\alias{get.hierarchy.hist}
\title{Converts the hierarchy.hist data set to a ddobj}
\usage{
get.hierarchy.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
hierarchy.data <- get.hierarchy.hist()

}
