% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.hierarchy.int}
\alias{get.hierarchy.int}
\title{Converts the hierarchy.int data set to a ddobj}
\usage{
get.hierarchy.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
hierarchy.data <- get.hierarchy.int()

}
