% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.horses.int}
\alias{get.horses.int}
\title{Converts the horses.int data set to a ddobj}
\usage{
get.horses.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
horse.data <- get.horses.int()

}
