% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.hospital.hist}
\alias{get.hospital.hist}
\title{Converts the hospital.hist data set to a ddobj}
\usage{
get.hospital.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
hospital.data <- get.hospital.hist()

}
