% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.iris.int}
\alias{get.iris.int}
\title{Converts the iris.int data set to a ddobj}
\usage{
get.iris.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
iris.data <- get.iris.int()

}
