% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.joggers.mix}
\alias{get.joggers.mix}
\title{Converts the joggers.mix data set to a ddobj}
\usage{
get.joggers.mix()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
joggers.data <- get.joggers.mix()

}
