% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.judge2.int}
\alias{get.judge2.int}
\title{Converts the judge2.int data set to a ddobj}
\usage{
get.judge2.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
judge.data <- get.judge2.int()

}
