% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.lackinfo.int}
\alias{get.lackinfo.int}
\title{Converts the lackinfo.int data set to a ddobj}
\usage{
get.lackinfo.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
lackinfo.data <- get.lackinfo.int()

}
