% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.lisbon_air_quality.int}
\alias{get.lisbon_air_quality.int}
\title{Converts the lisbon_air_quality.int data set to a ddobj}
\usage{
get.lisbon_air_quality.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
Lisbon.data <- get.lisbon_air_quality.int()

}
