% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.loans_by_risk.int}
\alias{get.loans_by_risk.int}
\title{Converts the loans_by_risk.int data set to a ddobj}
\usage{
get.loans_by_risk.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
loans.data <- get.loans_by_risk.int()

}
