% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.LoansbyPurpose_minmaxDt}
\alias{get.LoansbyPurpose_minmaxDt}
\title{Converts the LoansbyPurpose_minmaxDt data set to a ddobj}
\usage{
get.LoansbyPurpose_minmaxDt()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{MAINT.Data} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
Loan.data <- get.LoansbyPurpose_minmaxDt()

}
