% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.LoansbyRiskLvs_qntlDt}
\alias{get.LoansbyRiskLvs_qntlDt}
\title{Converts the LoansbyRiskLvs_qntlDt data set to a ddobj}
\usage{
get.LoansbyRiskLvs_qntlDt()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{MAINT.Data} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
Loan.data <- get.LoansbyRiskLvs_qntlDt()

}
