% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.lung_cancer.hist}
\alias{get.lung_cancer.hist}
\title{Converts the lung_cancer.hist data set to a ddobj}
\usage{
get.lung_cancer.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
lung_cancer.data <- get.lung_cancer.hist()

}
