% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.lynne1.int}
\alias{get.lynne1.int}
\title{Converts the lynne1.int data set to a ddobj}
\usage{
get.lynne1.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
Lynne.data <- get.lynne1.int()

}
