% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.mtcars.mix}
\alias{get.mtcars.mix}
\title{Converts the mtcars.mix data set to a ddobj}
\usage{
get.mtcars.mix()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
mtcars.data <- get.mtars.mix()

}
