% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.mushroom.int}
\alias{get.mushroom.int}
\title{Converts the mushroom.int data set to a ddobj}
\usage{
get.mushroom.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
mushroom.data <- get.mushroom.int()

}
