% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.mushroom_fuzzy.mix}
\alias{get.mushroom_fuzzy.mix}
\title{Converts the mushroom_fuzzy.mix data set to a ddobj}
\usage{
get.mushroom_fuzzy.mix()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
mushroom.data <- get.mushroom_fuzzy.mix()

}
