% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.nycflights.int}
\alias{get.nycflights.int}
\title{Converts the nycflights.int data set to a ddobj}
\usage{
get.nycflights.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
nycflights.data <- get.nycflights.int()

}
