% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.occupations.modal}
\alias{get.occupations.modal}
\title{Converts the occupations.modal data set to a ddobj}
\usage{
get.occupations.modal()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
occupations.data <- get.occupations.modal()

}
