% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.ohtemp.int}
\alias{get.ohtemp.int}
\title{Converts the ohtemp.int data set to a ddobj}
\usage{
get.ohtemp.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
ohtemp.data <- get.ohtemp.int()

}
