% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.oils.int}
\alias{get.oils.int}
\title{Converts the oils.int data set to a ddobj}
\usage{
get.oils.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
oils.data <- get.oils.int()

}
