% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.ozone.hist}
\alias{get.ozone.hist}
\title{Converts the ozone.hist data set to a ddobj}
\usage{
get.ozone.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
ozone.data <- get.ozone.hist()

}
