% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.OzoneFull}
\alias{get.OzoneFull}
\title{Converts the OzoneFull data set to a ddobj}
\usage{
get.OzoneFull()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{HistDAWass} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
blood.data <- get.OzoneFull()

}
