% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.polish_cars.mix}
\alias{get.polish_cars.mix}
\title{Converts the polish_cars.mix data set to a ddobj}
\usage{
get.polish_cars.mix()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
cars.data <- get.polish_cars.mix()

}
