% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.polish_voivodships.int}
\alias{get.polish_voivodships.int}
\title{Converts the polish_voivodships.int data set to a ddobj}
\usage{
get.polish_voivodships.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
Polish.data <- get.polish_voivodships.int()

}
