% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.profession.int}
\alias{get.profession.int}
\title{Converts the profession.int data set to a ddobj}
\usage{
get.profession.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
profession.data <- get.profession.int()

}
