% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.prostate.int}
\alias{get.prostate.int}
\title{Converts the prostate.int data set to a ddobj}
\usage{
get.prostate.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
prostate.data <- get.prostate.int()

}
