% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.simulated.hist}
\alias{get.simulated.hist}
\title{Converts the simulated.hist data set to a ddobj}
\usage{
get.simulated.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
simulated.data <- get.simulated.hist()

}
