% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.soccer.bivar.int}
\alias{get.soccer.bivar.int}
\title{Converts the soccer.bivar.int data set to a ddobj}
\usage{
get.soccer.bivar.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
soccer.data <- get.soccer.bivar.int()

}
