% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.state_income.hist}
\alias{get.state_income.hist}
\title{Converts the state_income.hist data set to a ddobj}
\usage{
get.state_income.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
income.data <- get.state_income.hist()

}
