% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.synthetic_clusters.int}
\alias{get.synthetic_clusters.int}
\title{Converts the synthetic_clusters.int data set to a ddobj}
\usage{
get.synthetic_clusters.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
cluster.data <- get.synthetic_clusters.int()

}
