% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.teams.int}
\alias{get.teams.int}
\title{Converts the teams.int data set to a ddobj}
\usage{
get.teams.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
teams.data <- get.teams.int()

}
