% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.temperature_city.int}
\alias{get.temperature_city.int}
\title{Converts the temperature_city.int data set to a ddobj}
\usage{
get.temperature_city.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
temperature.data <- get.temperature_city.int()

}
