% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.tennis.int}
\alias{get.tennis.int}
\title{Converts the tennis.int data set to a ddobj}
\usage{
get.tennis.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
tennis.data <- get.tennis.int()

}
