% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.town_services.mix}
\alias{get.town_services.mix}
\title{Converts the town_services.mix data set to a ddobj}
\usage{
get.town_services.mix()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
services.data <- get.town_services.mix()

}
