% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.trivial_intervals.int}
\alias{get.trivial_intervals.int}
\title{Converts the trivial_intervals.int data set to a ddobj}
\usage{
get.trivial_intervals.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
interval.data <- get.trivial_intervals.int()

}
