% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.uscrime.int}
\alias{get.uscrime.int}
\title{Converts the uscrime.int data set to a ddobj}
\usage{
get.uscrime.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
crime.data <- get.uscrime.int()

}
