% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.utsnow.int}
\alias{get.utsnow.int}
\title{Converts the utsnow.int data set to a ddobj}
\usage{
get.utsnow.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
snow.data <- get.utsnow.int()

}
