% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.veterinary.int}
\alias{get.veterinary.int}
\title{Converts the veterinary.int data set to a ddobj}
\usage{
get.veterinary.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
vet.data <- get.veterinary.int()

}
