% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.water_flow.int}
\alias{get.water_flow.int}
\title{Converts the water_flow.int data set to a ddobj}
\usage{
get.water_flow.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
flow.data <- get.water_flow.int()

}
