% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.weight_age.hist}
\alias{get.weight_age.hist}
\title{Converts the weight_age.hist data set to a ddobj}
\usage{
get.weight_age.hist()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
weight_age.data <- get.weight_age.hist()

}
