% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.wine.int}
\alias{get.wine.int}
\title{Converts the wine.int data set to a ddobj}
\usage{
get.wine.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
wine.data <- get.wine.int()

}
