% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.world_cup.int}
\alias{get.world_cup.int}
\title{Converts the world_cup.int data set to a ddobj}
\usage{
get.world_cup.int()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
world_cup.data <- get.world_cup.int()

}
