% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{histpdf}
\alias{histpdf}
\title{Probability density function of a histogram scaled data unit}
\usage{
histpdf(obj, i, j)
}
\arguments{
\item{obj}{an object of class \code{ddojb}}

\item{i}{unit number}

\item{j}{variable number}
}
\value{
a function(x) that will evaluate the p.d.f. at the values in x
}
\description{
Probability density function of a histogram scaled data unit
}
