% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{int.to.hist}
\alias{int.to.hist}
\title{Converts an interval scaled variable to a histogram scaled variable with a single bin and
proportion one}
\usage{
int.to.hist(x)
}
\arguments{
\item{x}{a list, typically a single component of an object of class \code{ddobj}}
}
\value{
a list, a single histogram scaled variable component for an object of class \code{ddobj}
}
\description{
Converts an interval scaled variable to a histogram scaled variable with a single bin and
proportion one
}
\examples{
obj <- suminto.ddobj (esoph, units = "agegp", interval="ncases", histogram="ncontrols")
int.to.hist (obj$ncases)

}
