% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{int_prod}
\alias{int_prod}
\title{Product of two quantile functions using integrate()}
\usage{
int_prod(x, i, j)
}
\arguments{
\item{x}{an object of class ddobj}

\item{i}{unit number}

\item{j}{variable number}
}
\value{
a scalar value
}
\description{
Product of two quantile functions using integrate()
}
