% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{intervals}
\alias{intervals}
\title{Format aesthetics for intervals}
\usage{
intervals (bp,  which = 1:bp$g, col = biplotEZ:::ez.col, lwd = 1, lty = 1,
label = FALSE, label.name = rownames(bp$X[[1]]$values), label.col=NULL, label.cex = 0.75,
label.side = "bottom", label.offset = 0.5)
}
\arguments{
\item{bp}{an object of class \code{ddbiplot}.}

\item{which}{a vector containing the groups or classes for which the intervals should be
displayed, with default \code{bp$g}.}

\item{col}{the colour(s) for the intervals, with default \code{blue}.}

\item{lwd}{the line width for the intervals, with default \code{1}.}

\item{lty}{the line type for the intervals, with default \code{1}.}

\item{label}{a logical value indicating whether the intervals should be labelled, with default
\code{FALSE}.}

\item{label.name}{a vector of the same length as \code{which} with label names for the intervals,
with default \code{NULL}. If \code{NULL}, the rownames of the intervals are
used. Alternatively, a custom vector of length \code{n} should be used.}

\item{label.col}{a vector of the same length as \code{which} with label colours for the intervals,
with default as the same colour of the intervals.}

\item{label.cex}{a vector of the same length as \code{which} with label text expansions for the
intervals, with default \code{0.75}.}

\item{label.side}{the side at which the label of the interval appears, with default
\code{bottom}.  Note that unlike the argument \code{pos} in \code{text()},
options are not \code{1}, \code{2}, \code{3}, \code{4}, but "\code{bottom}",
"\code{left}", "\code{top}", "\code{right}" and in addition "\code{mid.bottom}",
"\code{mid.left}", "\code{mid.right}", "\code{mid.top}", "\code{bottom.left}",
"\code{bottom.right}", "\code{top.left}", and "\code{top.right}" are also allowed.
The latter four options only take effect when the \code{type} argument in \code{plot}
is "\code{diagonal}".}

\item{label.offset}{the offset of the label from the interval. See \code{?text} for a
detailed explanation of the argument \code{offset} when \code{label.side} is one of
"\code{bottom}", "\code{left}", "\code{top}", "\code{right}". For the other options
of \code{label.side}, the \code{label.offset} values is applied as \code{offset}-0.5
units in the plot.}
}
\value{
The object of class \code{ddbiplot} will be appended with a list called \code{intervals}
        containing the following elements:
\item{which}{a vector containing the groups or classes for which the samples (and means) are
             displayed.}
\item{col}{the colour(s) of the intervals.}
\item{lwd}{the line width of the intervals.}
\item{lty}{the line type of the intervals.}
\item{label}{a logical value indicating whether intervals are labelled.}
\item{label.name}{the label names of the samples.}
\item{label.col}{the label colours of the samples.}
\item{label.cex}{the label text expansions of the samples.}
\item{label.side}{the side at which the label of the interval appears.}
\item{label.offset}{the offset of the label from the plotted intervals.}
}
\description{
This function allows the user to format the aesthetics for the units when
units are displayed as \code{type = "intervals"} or \code{type = "diagonal"}
in \code{plot()}.
}
\examples{
ddbiplot(data = Oils.data) |> PCA() |> intervals(col="purple",lwd=2, label = TRUE) |> plot()

}
