% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{intpdf}
\alias{intpdf}
\title{Probability density function of a interval scaled data unit}
\usage{
intpdf(obj, i, j)
}
\arguments{
\item{obj}{an object of class \code{ddojb}}

\item{i}{unit number}

\item{j}{variable number}
}
\value{
a function(x) that will evaluate the p.d.f. at the values in x
}
\description{
Probability density function of a interval scaled data unit
}
\details{
This function also operates on numeric data, viewed as zero length intervals.
In theory the density is infinite at the numeric value. For computational
purposes, the density value is replaced by twice the largest density value
of the other units on the same variable.
}
