% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{matH.to.data.frame}
\alias{matH.to.data.frame}
\title{Converts a matH object (package HistDAWass) to a data.frame}
\usage{
matH.to.data.frame(obj)
}
\arguments{
\item{obj}{matH obj}
}
\value{
A list with the following components is available:
\item{df}{the data.frame constructed from the matH object.}
\item{cols}{a vector with the first column number for each variable.}
\item{n.int}{a vector with the number of intervals for each variable.}
}
\description{
Converts a matH object (package HistDAWass) to a data.frame
}
