% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{modal_extract}
\alias{modal_extract}
\title{Extract modal variables from the dataSDA package}
\usage{
modal_extract(x, normalize = TRUE)
}
\arguments{
\item{x}{modal data from a data set in dataSDA}

\item{normalize}{logical value to ensure probabilities add up to 1}
}
\value{
a data frame for \code{create_ddobj()}
}
\description{
Extract modal variables from the dataSDA package
}
