% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{num.to.int}
\alias{num.to.int}
\title{Converts a numeric variable to an interval scaled variable with zero interval width}
\usage{
num.to.int(x)
}
\arguments{
\item{x}{a list, typically a single component of an object of class \code{ddobj}}
}
\value{
a list, a single interval scaled variable component for an object of class \code{ddobj}
}
\description{
Converts a numeric variable to an interval scaled variable with zero interval width
}
\examples{
num.to.int (Oils.data$SO.LDP)

}
