% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA.ddbiplot}
\alias{PCA.ddbiplot}
\title{Calculate elements for the PCA biplot for distributional data}
\usage{
\method{PCA}{ddbiplot}(
  bp,
  dim.biplot = c(2, 1, 3),
  e.vects = 1:bp$p,
  group.aes = NULL,
  show.class.means = FALSE,
  correlation.biplot = FALSE,
  ...
)
}
\arguments{
\item{bp}{an object of class \code{ddbiplot} obtained from preceding function \code{ddbiplot()}.}

\item{dim.biplot}{the dimension of the biplot. Only values \code{1}, \code{2} and \code{3} are accepted, with default \code{2}.}

\item{e.vects}{the vector indicating which eigenvectors (principal components) should be plotted in the biplot, with default \code{1:dim.biplot}.}

\item{group.aes}{a vector of the same length as the number of rows in the data matrix
for differentiated aesthetics for samples.}

\item{show.class.means}{a logical value indicating whether group means should be plotted in the
biplot.}

\item{correlation.biplot}{a logical value. If \code{FALSE}, the distances between sample points are
optimally approximated in the biplot. If \code{TRUE}, the correlations between
variables are optimally approximated by the cosine of the angles between
axes. Default is \code{FALSE}.}

\item{...}{additional arguments}
}
\value{
an object of class \code{ddPCA}, inherits from class \code{ddbiplot}.
}
\description{
This function performs calculations for the construction of a PCA biplot for
             distributional data.
}
\examples{
ddbiplot(data = Oils.data) |> PCA() |> plot()

}
