% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{prediction}
\alias{prediction}
\title{Predict biplot samples}
\usage{
prediction(bp, samples = TRUE, type = "intervals")
}
\arguments{
\item{bp}{an object of class \code{ddbiplot} obtained from preceding function \code{ddbiplot()}.}

\item{samples}{logical TRUE to predict samples; which sample numbers to be implemented}

\item{type}{either default \code{intervals} or \code{vertices}. Which unit
representation to predict}
}
\value{
an object of class \code{ddbiplot} with additional component
         \code{predict$samples}
}
\description{
Predict biplot samples
}
\examples{
ddbiplot(data = Oils.data) |> PCA() |>
  prediction (type = "intervals") |> plot(type = "intervals")

}
