% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{print.ddobj}
\alias{print.ddobj}
\title{Generic print function for objects of class ddobj}
\usage{
\method{print}{ddobj}(x, ...)
}
\arguments{
\item{x}{an object of class \code{ddobj}.}

\item{...}{additional arguments.}
}
\value{
This function will not produce a return value, it is called for side effects.
}
\description{
This function is used to print a distributional data object.
}
\examples{
my.obj <- create_ddobj (toy.data, type=c("numeric","interval","histogram","categorical", "modal"),
          cols = c(1, 2, 4, 13, 14), n.int=4, n.cat=3)
print (my.obj)

}
