% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{quantij}
\alias{quantij}
\title{Computes the quantile function}
\usage{
quantij(obj, i, j)
}
\arguments{
\item{obj}{an object of class \code{ddojb}}

\item{i}{unit number}

\item{j}{variable number}
}
\value{
a function(tvec) that will evaluate the inverse of the c.d.f. at the values
in tvec, where 0 <= tvec <= 1
}
\description{
Computes the quantile function
}
