% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{sqL2Wass_dist}
\alias{sqL2Wass_dist}
\title{Computes the squared L2 Wasserstein distance}
\usage{
sqL2Wass_dist(obj)
}
\arguments{
\item{obj}{an object of class \code{ddobj}}
}
\value{
a symmetric matrix of squared distances
}
\description{
Computes the squared L2 Wasserstein distance
}
\examples{
obj <- suminto_ddobj (esoph, units = "agegp", interval="ncases", histogram="ncontrols")
sqL2Wass_dist(obj)

}
