% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{suminto_ddobj}
\alias{suminto_ddobj}
\title{Summarise a data set into a distributional data object}
\usage{
suminto_ddobj(df, units = names(df)[1],
                     interval=NULL, histogram=NULL, modal=NULL)
}
\arguments{
\item{df}{data frame from which the \code{ddobj} object is created}

\item{units}{the column names that will form the units}

\item{interval}{the column names that will be summarised into interval scaled data}

\item{histogram}{the column names that will be summarised into histogram scaled data}

\item{modal}{the column names that will be summarised into modal data}
}
\value{
and object of class \code{ddobj}
}
\description{
Summarise a data set into a distributional data object
}
\examples{
suminto_ddobj (esoph, units = "agegp", interval="ncases",
               histogram="ncontrols", modal=c("alcgp","tobgp"))

}
