% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{toy.data}
\alias{toy.data}
\title{Toy data of illustrating different types of distributional and 'ordinary' data.}
\format{
A data frame with 5 rows and 19 columns:
\describe{
  \item{X1}{numeric variable}
  \item{X2}{first endpoint of the interval scaled variable}
  \item{X2.up}{second endpoint of the interval scaled variable. Typically the first endpoint is
               the lower endpoint and the second the upper endpoint, however, if specified in
               reverse, the function \code{create_ddobj} will switch the endpoints}
  \item{X3}{lower end of the first interval for the histogram scaled variable}
  \item{X3.I1}{upper end of the first interval for the histogram scaled variable}
  \item{X3.I2}{upper end of the second interval for the histogram scaled variable}
  \item{X3.I3}{upper end of the third interval for the histogram scaled variable}
  \item{X3.I4}{upper end of the fourth interval for the histogram scaled variable}
  \item{X3.p1}{proportion of observations in the first interval for the histogram scaled variable}
  \item{X3.p2}{proportion of observations in the second interval for the histogram scaled variable}
  \item{X3.p3}{proportion of observations in the third interval for the histogram scaled variable}
  \item{X3.p4}{proportion of observations in the fourth interval for the histogram scaled variable}
  \item{X4}{categorical variable}
  \item{X5}{first category for the modal variable}
  \item{X5.c2}{second category for the modal variable}
  \item{X5.c3}{third category for the modal variable}
  \item{X5.p1}{proportion of observations in the first category for the modal variable}
  \item{X5.p2}{proportion of observations in the second category for the modal variable}
  \item{X5.p3}{proportion of observations in the third category for the modal variable}
  }

  "toy.data"
}
\usage{
toy.data
}
\description{
A data set containing a numeric, an interval scaled, a histogram scaled, a categorical
and a modal variable.
}
