% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{uniformly_dense_intervals}
\alias{uniformly_dense_intervals}
\title{Computes uniformly dense intervals}
\usage{
uniformly_dense_intervals(x)
}
\arguments{
\item{x}{an object of class \code{ddobj}}
}
\value{
a list with the following components
\item{dens.int}{an nxpxs array with the s values defining the (s-1) uniformly dense
intervals for each of the nxp unit-variable combinations}
\item{proportions}{the vector of uniform proportions}
}
\description{
Computes uniformly dense intervals
}
