% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{vertices}
\alias{vertices}
\title{Format aesthetics for vertices representation}
\usage{
vertices (bp,  which = 1:bp$g, col = biplotEZ:::ez.col, pch = 1, cex = 0.3,
          lwd = 1, lty = 1, label = FALSE, label.name = rownames(bp$X[[1]]$values),
          label.col=NULL, label.cex = 0.75, label.side = "bottom", label.offset = 0.5,
          type = "convexhull")
}
\arguments{
\item{bp}{an object of class \code{ddbiplot}.}

\item{which}{a vector containing the groups or classes for which the vertices should be
displayed, with default \code{bp$g}.}

\item{col}{the colour(s) for the vertices, with default \code{blue}.}

\item{pch}{the plotting character(s) for the vertices, with default \code{1}.}

\item{cex}{the character expansion of the plotting character(s), with default \code{1}.}

\item{lwd}{the line width for the vertices, with default \code{1}.}

\item{lty}{the line type for the vertices, with default \code{1}.}

\item{label}{a logical value indicating whether the vertices should be labelled, with default
\code{FALSE}.}

\item{label.name}{a vector of the same length as \code{which} with label names for the vertices,
with default \code{NULL}. If \code{NULL}, the rownames of the units are
used. Alternatively, a custom vector of length \code{n} should be used.}

\item{label.col}{a vector of the same length as \code{which} with label colours for the vertices,
with default as the same colour of the vertices.}

\item{label.cex}{a vector of the same length as \code{which} with label text expansions for the
vertices, with default \code{0.3}.}

\item{label.side}{the side at which the label of the unit appears, with default
\code{bottom}. Note that unlike the argument \code{pos} in \code{text()},
options are "\code{bottom}", "\code{left}", "\code{top}", "\code{right}" and
not \code{1}, \code{2}, \code{3}, \code{4}.}

\item{label.offset}{the offset of the label from the plotted vertices. See \code{?text} for a
detailed explanation of the argument \code{offset}.}

\item{type}{either \code{"connect"} to connect vertices in full space such as a cube in 3D or
\code{"convexhull"} to construct a convex hull around the vertices.}
}
\value{
The object of class \code{ddbiplot} will be appended with a list called \code{vertices}
        containing the following elements:
\item{which}{a vector containing the groups or classes for which the samples (and means) are
             displayed.}
\item{col}{the colour(s) of the vertices.}
\item{pch}{the plotting character(s) of the vertices.}
\item{cex}{the character expansion of the plotting character(s).}
\item{lwd}{the line width for connecting vertices or a convex hull.}
\item{lty}{the line type for connecting vertices or a convex hull.}
\item{label}{a logical value indicating whether vertices are labelled.}
\item{label.name}{the label names of the vertices.}
\item{label.col}{the label colours of the vertices.}
\item{label.cex}{the label text expansions of the vertices.}
\item{label.side}{the side at which the label of the vertices appears..}
\item{label.offset}{the offset of the label from the vertices.}
\item{type}{either \code{"convexhull"} or \code{"connect"}}
}
\description{
This function allows the user to format the aesthetics for the units.
}
\examples{
ddbiplot(data = Oils.data) |> PCA() |> vertices(col="purple",lwd=2, label = TRUE) |> plot()

}
