% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{WassL2var}
\alias{WassL2var}
\title{Computes the L2 Wasserstein variance or covariance of (a) distributional data object(s)}
\usage{
WassL2var(x)
}
\arguments{
\item{x}{an object of class \code{ddojb}}
}
\value{
variance-covariance matrix
}
\description{
This function uses the \code{WH.var.covar} method from the \code{HistDAWass} package to
compute variances and covariances. Interval scaled data are converted to histogram
scaled with a single bin and proportion = 1.
}
\examples{
obj <- suminto.ddobj (esoph, units = "agegp", interval="ncases", histogram="ncontrols")
WassL2var (obj)

}
\references{
Irpino, A. and Verde, R. 2015. Basic statistics for distributional symbolic variables:
a new metric-based approach. Advances in Data Analysis and Classification, 9(2), pp.143-157.
}
