% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{create.vertices}
\alias{create.vertices}
\title{Create a vertices matrix from interval scaled data}
\usage{
create.vertices(obj, i, connect = FALSE)
}
\arguments{
\item{obj}{an object of class \code{ddobj}}

\item{i}{number of the entity for which the vertices matrix is computed}

\item{connect}{logical argument indicating whether connections between vertices should be
computed. Note this requires evaluating (2^p) chose 2 possible connections}
}
\value{
a list with two components:
\item{vertices}{a matrix of size \eqn{2^p \times p} where \eqn{p} is the number of intervals.}
\item{connections}{a two-column matrix with each row indicating the
                   numbers of two rows of \code{vertices} to be
                   connected such as a cube in 3D.}
}
\description{
Create a vertices matrix from interval scaled data
}
\examples{
obj <- suminto.ddobj (esoph, entities = "agegp", interval=c("ncases","ncontrols"))
create.vertices (obj, 1)

}
