% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{ddcormat}
\alias{ddcormat}
\title{Computes the correlation matrix of distributional data variables}
\usage{
ddcormat(obj)
}
\arguments{
\item{obj}{an object of class \code{ddobj}}
}
\value{
a correlation matrix of size length of obj by length of obj
}
\description{
Computes the correlation matrix of distributional data variables
}
\examples{
obj <- suminto.ddobj (esoph, entities = "agegp", interval="ncases", histogram="ncontrols")
ddcormat (obj)
}
