% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{ddcov}
\alias{ddcov}
\title{Computes the covariance of two distributional data objects}
\usage{
ddcov(x, y)
}
\arguments{
\item{x}{a list, typically a single component of an object of class \code{ddobj}}

\item{y}{a list, typically a single component of an object of class \code{ddobj}}
}
\value{
numeric variance or covariance value
}
\description{
This function only works for an interval scaled or histogram scaled variable. For an interval
scaled variable \code{x} should be a list with component \code{$type = "interval"} and
component \code{values} a two-column matrix of interval endpoints. If \code{x} is a histogram
scaled variable, it should be a list with component \code{$type = "histogram"} and components
\code{intervals} and \code{proportions}.
}
\examples{
obj <- suminto.ddobj (esoph, entities = "agegp", interval="ncases", histogram="ncontrols")
ddcov (obj$ncases, obj$ncontrols)

}
\references{
Billard, L. and Diday, E. (2003) From the Statistics of Data to the Statistics of Knowledge.
\emph{Journal of the American Statistical Association.} 98:470-487.
}
