% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{ddvar}
\alias{ddvar}
\title{Computes the variance or covariance of (a) distributional data object(s)}
\usage{
ddvar(x, y)
}
\arguments{
\item{x}{a list, typically a single component of an object of class \code{ddojb}}

\item{y}{optional, a list, typically a single component of an object of class \code{ddojb}. If
\code{y} is specified, the covariance is computed.}
}
\value{
numeric variance or covariance value
}
\description{
This function only works for an interval scaled or histogram scaled variable. For an interval
scaled variable \code{x} should be a list with component \code{$type = "interval"} and
component \code{values} a two-column matrix of interval endpoints. If \code{x} is a histogram
scaled variable, it should be a list with component \code{$type = "histogram"} and components
\code{intervals} and \code{proportions}.
}
\examples{
obj <- suminto.ddobj (esoph, entities = "agegp", interval="ncases", histogram="ncontrols")
ddvar (obj$ncases) # variance
ddvar (obj$ncases, obj$ncontrols) # covariance

}
\references{
Billard, L. and Diday, E. (2003) From the Statistics of Data to the Statistics of Knowledge.
\emph{Journal of the American Statistical Association.} 98:470-487.
}
