% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{get.Face.iGAP}
\alias{get.Face.iGAP}
\title{Converts the Face.iGAP data set to a ddobj}
\usage{
get.Face.iGAP()
}
\value{
an object of class \code{ddobj}
}
\description{
Obtain the data set from the package \code{dataSDA} and convert is into a \code{ddobj}
for use in \code{ddbiplot()}.
}
\examples{
face.data <- get.Face.iGAP()

}
