% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{Oils.data}
\alias{Oils.data}
\title{Oils data of Ichino. 1988. General metrics for mixed features the Cartesian space theory for
              pattern recognition. In Proceedings of the 1988 IEEE International Conference on
              Systems, Man, and Cybernetics (Vol. 1, pp. 494-497). IEEE.}
\format{
An object of class \code{ddobj}. A list of length 5:
\describe{
  \item{Spec.gravity}{interval scaled variable}
  \item{Freezing.point}{interval scaled variable}
  \item{Iodine.value}{interval scaled variable}
  \item{Saponification}{interval scaled variable}
  \item{SO.LDP}{numeric variable}
  }

  "Oils.data"
}
\usage{
Oils.data
}
\description{
A data set containing 8 entities with four interval scaled variables and one numeric variable.
}
\keyword{datasets}
