% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{sq.L2.Wass_dist}
\alias{sq.L2.Wass_dist}
\title{Computes the squared l2 Wasserstein distance}
\usage{
sq.L2.Wass_dist(obj)
}
\arguments{
\item{obj}{an object of class \code{ddobj}}
}
\value{
a symmetric matrix of squared distances
}
\description{
Computes the squared l2 Wasserstein distance
}
\examples{
obj <- suminto.ddobj (esoph, entities = "agegp", interval="ncases", histogram="ncontrols")
sq.L2.Wass_dist(obj)

}
